requirements.txt
pandas==1.5.3
inflection==0.5.1
plotly==5.9.0
folium==0.17.0
streamlit==1.37.1
Pillow==9.4.0
streamlit_folium==0.22.0